requirements.txt
scipy==1.2.1
scikit-image==0.17.1
tensorflow==1.15.0
opencv-python==4.5.1.48
torchvision
pyyaml
jupyterlab
pillow
glog
easydict
matplotlib
numpy
google-cloud-vision
azure-cognitiveservices-vision-computervision
pylint
